[refactor] Refactoring constants, types, etc.
