+++ src/apis/card.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { ICardData } from '@/types/apis';
+import { ICardData } from '@/types/apisTypes';
 
 export const getAllCards = async (): Promise<ICardData[]> => {
   const response = await axios.get<ICardData[]>('/card/front/all');
