+++ src/components/common/Error.tsx
@@ -1,0 +1,5 @@
+export const Error = () => {
+    return(
+        <div>에러</div>
+    )
+}
