+++ src/pages/main/MainSelect.tsx
@@ -1,9 +1,10 @@
-import { ThreeCard } from '../../components/common/ThreeCard';
+import { ThreeCard } from '@/components/game/ThreeCard';
 import { useLocation } from 'react-router-dom';
+import { DEFAULT_QUESTION } from '@/utils/constants';
 
 export const MainSelect = () => {
   const location = useLocation();
-  const question = location.state.gameData.game_question || '무슨 카드가 제일 끌리나요?';
+  const question = location.state.gameData.game_question || DEFAULT_QUESTION;
   const randomCards = location.state.randomCards
 
   return (
