+++ src/utils/constants.ts
@@ -1,0 +1,39 @@
+//최대 선택 카드 갯수
+export const MAX_SELECTED_CARD = 3;
+//카드 갯수
+export const CARD_NUM = 3;
+//기본 질문
+export const DEFAULT_QUESTION = '무슨 카드가 제일 끌리나요?'
+//기본 랜덤 카드 3개
+export const DEFAULT_RANDOM_THREE_CARDS = [
+    {
+      "card_id": "2825f9c0-c482-4925-bfb4-ac963590e307",
+      "card_name": "The Fool",
+      "card_number": 1,
+      "card_image_url": "https://upload.wikimedia.org/wikipedia/commons/9/90/RWS_Tarot_00_Fool.jpg",
+      "card_contents": {
+          "forward": "The Fool card represents new beginnings, innocence, and a leap of faith. It signifies embarking on a journey with an open heart and a sense of adventure. This card encourages you to trust in the universe and embrace the unknown.",
+          "reverse": "The reverse of The Fool card may indicate recklessness, naivety, or missed opportunities. It may suggest a need to be more cautious and grounded in your approach to life."
+      }
+  },
+  {
+      "card_id": "d5b4777d-c038-4bd7-9d41-78c3b8bfc9b4",
+      "card_name": "The Magician",
+      "card_number": 2,
+      "card_image_url": "https://upload.wikimedia.org/wikipedia/commons/d/de/RWS_Tarot_01_Magician.jpg",
+      "card_contents": {
+          "forward": "The Magician card represents manifestation, resourcefulness, and power. It signifies the ability to harness your skills and talents to achieve your goals. This card encourages you to take action and make things happen through focused intention and willpower.",
+          "reverse": "The reverse of The Magician card may indicate manipulation, untapped potential, or misuse of power. It may suggest a need to be mindful of your actions and their consequences."
+      }
+  },
+  {
+      "card_id": "2a5f63c2-b74c-44ee-95f2-ff975ffadf67",
+      "card_name": "The High Priestess",
+      "card_number": 3,
+      "card_image_url": "https://upload.wikimedia.org/wikipedia/commons/8/88/RWS_Tarot_02_High_Priestess.jpg",
+      "card_contents": {
+          "forward": "The High Priestess card represents intuition, mystery, and inner wisdom. It signifies a deep connection to the subconscious mind and the mysteries of the universe. This card encourages you to listen to your inner voice and trust your instincts.",
+          "reverse": "The reverse of The High Priestess card may indicate hidden agendas, secrets, or deception. It may suggest a need to look beneath the surface and uncover the truth in a situation."
+      }
+  }
+  ]
